Move files and update libraries
